Rezervacija karti.

+++ server-app/src/main/java/isa/project/repository/aircompany/TicketRespository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import isa.project.model.aircompany.Ticket;
 
-public interface TicketRespository extends JpaRepository<Ticket, Integer> {
+public interface TicketRespository extends JpaRepository<Ticket, Long> {
 
 }
